+++ tests/utils/generate_pq.cpp
@@ -2,7 +2,9 @@
 // Licensed under the MIT license.
 
 #include "math_utils.h"
-#include "partition_and_pq.h"
+#include "pq.h"
+#include "partition.h"
+
 
 #define KMEANS_ITERS_FOR_PQ 15
 
@@ -10,7 +12,7 @@ template<typename T>
 bool generate_pq(const std::string& data_path,
                  const std::string& index_prefix_path,
                  const size_t num_pq_centers, const size_t num_pq_chunks,
-                 const float sampling_rate) {
+                 const float sampling_rate, const bool opq) {
   std::string pq_pivots_path = index_prefix_path + "_pq_pivots.bin";
   std::string pq_compressed_vectors_path =
       index_prefix_path + "_pq_compressed.bin";
@@ -23,13 +25,19 @@ bool generate_pq(const std::string& data_path,
   std::cout << "For computing pivots, loaded sample data of size " << train_size
             << std::endl;
 
-  //  generate_pq_pivots(train_data, train_size, train_dim, num_pq_centers,
-  //                     num_pq_chunks, KMEANS_ITERS_FOR_PQ, pq_pivots_path);
-  generate_opq_pivots(train_data, train_size, train_dim, num_pq_centers,
-                      num_pq_chunks, pq_pivots_path, true);
-  generate_pq_data_from_pivots<T>(data_path, num_pq_centers, num_pq_chunks,
-                                  pq_pivots_path, pq_compressed_vectors_path,
-                                  true);
+  if (opq) {
+    diskann::generate_opq_pivots(train_data, train_size, train_dim,
+                                 num_pq_centers, num_pq_chunks, pq_pivots_path,
+                                 true);
+
+  } else {
+    diskann::generate_pq_pivots(train_data, train_size, train_dim,
+                                num_pq_centers, num_pq_chunks,
+                                KMEANS_ITERS_FOR_PQ, pq_pivots_path);
+  }
+  diskann::generate_pq_data_from_pivots<T>(data_path, num_pq_centers,
+                                           num_pq_chunks, pq_pivots_path,
+                                           pq_compressed_vectors_path, true);
 
   delete[] train_data;
 
@@ -37,12 +45,12 @@ bool generate_pq(const std::string& data_path,
 }
 
 int main(int argc, char** argv) {
-  if (argc != 6) {
+  if (argc != 7) {
     std::cout
         << "Usage: \n"
         << argv[0]
         << "  <data_type[float/uint8/int8]>   <data_file[.bin]>"
-           "  <PQ_prefix_path>  <target-bytes/data-point>  <sampling_rate>"
+           "  <PQ_prefix_path>  <target-bytes/data-point>  <sampling_rate> <PQ(0)/OPQ(1)>"
         << std::endl;
   } else {
     const std::string data_path(argv[2]);
@@ -50,16 +58,17 @@ int main(int argc, char** argv) {
     const size_t      num_pq_centers = 256;
     const size_t      num_pq_chunks = (size_t) atoi(argv[4]);
     const float       sampling_rate = atof(argv[5]);
+    const bool        opq = atoi(argv[6]) == 0 ? false : true; 
 
     if (std::string(argv[1]) == std::string("float"))
       generate_pq<float>(data_path, index_prefix_path, num_pq_centers,
-                         num_pq_chunks, sampling_rate);
+                         num_pq_chunks, sampling_rate, opq);
     else if (std::string(argv[1]) == std::string("int8"))
       generate_pq<int8_t>(data_path, index_prefix_path, num_pq_centers,
-                          num_pq_chunks, sampling_rate);
+                          num_pq_chunks, sampling_rate, opq);
     else if (std::string(argv[1]) == std::string("uint8"))
       generate_pq<uint8_t>(data_path, index_prefix_path, num_pq_centers,
-                           num_pq_chunks, sampling_rate);
+                           num_pq_chunks, sampling_rate, opq);
     else
       std::cout << "Error. wrong file type" << std::endl;
   }
