+++ tests/utils/partition_with_ram_budget.cpp
@@ -4,7 +4,7 @@
 #include <index.h>
 #include <math_utils.h>
 #include "cached_io.h"
-#include "partition_and_pq.h"
+#include "partition.h"
 
 // DEPRECATED: NEED TO REPROGRAM
